requirements.dev.txt
django==5.1.6
djangorestframework==3.15.2
django-money==3.5.3
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
django-silk
djangorestframework-simplejwt==5.4.0